各モジュールに移動できるファイルを移動する
