+++ app/src/main/java/com/github/kitakkun/foos/data/model/UserWithFollowState.kt
@@ -1,5 +1,6 @@
 package com.github.kitakkun.foos.data.model
 
+import com.github.kitakkun.foos.user.followlist.FollowState
 import com.github.kitakkun.foos.user.model.DatabaseUser
 
 data class UserWithFollowState(
