+++ app/src/main/java/com/github/kitakkun/foos/di/RepositoryModule.kt
@@ -4,6 +4,8 @@ import com.github.kitakkun.foos.data.repository.*
 import com.github.kitakkun.foos.di.FirebaseModule.provideFireStoreInstance
 import com.github.kitakkun.foos.di.FirebaseModule.provideFirebaseAuthInstance
 import com.github.kitakkun.foos.di.FirebaseModule.provideFirebaseStorageInstance
+import com.github.kitakkun.foos.user.repository.FollowRepository
+import com.github.kitakkun.foos.user.repository.FollowRepositoryImpl
 import com.github.kitakkun.foos.user.repository.UsersRepository
 import com.github.kitakkun.foos.user.repository.UsersRepositoryImpl
 import dagger.Module
