+++ app/src/main/java/com/github/kitakkun/foos/ui/view/screen/ScreenNavHost.kt
@@ -9,15 +9,13 @@ import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.navigation
+import com.github.kitakkun.foos.common.navigation.MainScreen
+import com.github.kitakkun.foos.common.navigation.StringList
+import com.github.kitakkun.foos.common.navigation.SubScreen
 import com.github.kitakkun.foos.ui.navigation.BottomSheet
-import com.github.kitakkun.foos.ui.navigation.MainScreen
-import com.github.kitakkun.foos.ui.navigation.SubScreen
-import com.github.kitakkun.foos.ui.navigation.navargs.StringList
 import com.github.kitakkun.foos.ui.view.bottomsheet.PostOptionBottomSheet
 import com.github.kitakkun.foos.ui.view.screen.auth.signin.SignInScreen
 import com.github.kitakkun.foos.ui.view.screen.auth.signup.SignUpScreen
-import com.github.kitakkun.foos.ui.view.screen.followlist.FollowListScreen
-import com.github.kitakkun.foos.ui.view.screen.followlist.FollowListViewModelImpl
 import com.github.kitakkun.foos.ui.view.screen.home.HomeScreen
 import com.github.kitakkun.foos.ui.view.screen.home.HomeViewModelImpl
 import com.github.kitakkun.foos.ui.view.screen.imagedetail.ImageDetailScreen
@@ -37,6 +35,8 @@ import com.github.kitakkun.foos.ui.view.screen.setting.SettingScreen
 import com.github.kitakkun.foos.ui.view.screen.setting.SettingViewModelImpl
 import com.github.kitakkun.foos.ui.view.screen.userprofile.UserProfileScreen
 import com.github.kitakkun.foos.ui.view.screen.userprofile.UserProfileViewModelImpl
+import com.github.kitakkun.foos.user.followlist.FollowListScreen
+import com.github.kitakkun.foos.user.followlist.FollowListViewModelImpl
 import com.google.accompanist.navigation.material.ExperimentalMaterialNavigationApi
 import com.google.accompanist.navigation.material.bottomSheet
