+++ app/src/main/java/com/github/kitakkun/foos/ui/view/screen/auth/signin/SignInViewModel.kt
@@ -7,7 +7,7 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.viewModelScope
 import com.github.kitakkun.foos.R
-import com.github.kitakkun.foos.ui.navigation.SubScreen
+import com.github.kitakkun.foos.common.navigation.SubScreen
 import com.google.firebase.auth.FirebaseAuth
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
