+++ app/src/main/java/com/github/kitakkun/foos/ui/view/screen/auth/signup/SignUpViewModel.kt
@@ -5,8 +5,8 @@ import androidx.compose.runtime.State
 import androidx.compose.runtime.mutableStateOf
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.github.kitakkun.foos.ui.navigation.MainScreen
-import com.github.kitakkun.foos.ui.navigation.SubScreen
+import com.github.kitakkun.foos.common.navigation.MainScreen
+import com.github.kitakkun.foos.common.navigation.SubScreen
 import com.github.kitakkun.foos.user.model.Email
 import com.github.kitakkun.foos.user.model.Password
 import com.github.kitakkun.foos.user.repository.UsersRepository
