+++ app/src/main/java/com/github/kitakkun/foos/ui/view/screen/locationselect/LocationSelectViewModelImpl.kt
@@ -4,7 +4,7 @@ import androidx.compose.runtime.State
 import androidx.compose.runtime.mutableStateOf
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.github.kitakkun.foos.ui.navigation.SubScreen
+import com.github.kitakkun.foos.common.navigation.SubScreen
 import com.github.kitakkun.foos.ui.state.screen.locationselect.LocationSelectScreenUiState
 import com.google.android.gms.maps.model.LatLng
 import dagger.hilt.android.lifecycle.HiltViewModel
