+++ app/src/main/java/com/github/kitakkun/foos/ui/view/screen/userprofile/UserProfileScreen.kt
@@ -15,13 +15,13 @@ import androidx.navigation.NavController
 import com.github.kitakkun.foos.R
 import com.github.kitakkun.foos.common.const.paddingMedium
 import com.github.kitakkun.foos.customview.preview.PreviewContainer
-import com.github.kitakkun.foos.ui.view.component.FollowInfo
 import com.github.kitakkun.foos.ui.view.component.MaxSizeLoadingIndicator
-import com.github.kitakkun.foos.ui.view.component.UserIcon
-import com.github.kitakkun.foos.ui.view.component.VerticalUserIdentityText
-import com.github.kitakkun.foos.ui.view.component.button.FollowButton
 import com.github.kitakkun.foos.ui.view.component.list.MediaPostGrid
 import com.github.kitakkun.foos.ui.view.component.list.PostItemList
+import com.github.kitakkun.foos.user.FollowButton
+import com.github.kitakkun.foos.user.FollowInfo
+import com.github.kitakkun.foos.user.UserIcon
+import com.github.kitakkun.foos.user.VerticalUserIdentityText
 import com.google.accompanist.pager.*
 import com.google.accompanist.swiperefresh.SwipeRefresh
 import com.google.accompanist.swiperefresh.rememberSwipeRefreshState
