+++ app/src/main/java/com/github/kitakkun/foos/ui/view/screen/userprofile/UserProfileViewModelImpl.kt
@@ -6,15 +6,15 @@ import androidx.compose.runtime.State
 import androidx.compose.runtime.mutableStateOf
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import com.github.kitakkun.foos.common.navigation.StringList
+import com.github.kitakkun.foos.common.navigation.SubScreen
 import com.github.kitakkun.foos.data.domain.fetcher.post.FetchPostsByUserIdUseCase
 import com.github.kitakkun.foos.data.domain.fetcher.post.FetchPostsUserReactedByUserIdUseCase
 import com.github.kitakkun.foos.data.domain.fetcher.post.FetchPostsWithMediaByUserIdUseCase
-import com.github.kitakkun.foos.data.repository.FollowRepository
 import com.github.kitakkun.foos.ui.navigation.BottomSheet
-import com.github.kitakkun.foos.ui.navigation.SubScreen
-import com.github.kitakkun.foos.ui.navigation.navargs.StringList
 import com.github.kitakkun.foos.ui.state.component.PostItemUiState
 import com.github.kitakkun.foos.ui.state.screen.userprofile.UserProfileScreenUiState
+import com.github.kitakkun.foos.user.repository.FollowRepository
 import com.github.kitakkun.foos.user.repository.UsersRepository
 import com.google.firebase.auth.ktx.auth
 import com.google.firebase.ktx.Firebase
