+++ common/src/main/res/drawable/ic_favorite.xml
