+++ common/src/main/res/drawable/ic_home.xml
