+++ common/src/main/res/drawable/ic_settings.xml
