+++ customview/src/main/java/com/github/kitakkun/foos/customview/composable/button/RoundIconActionButton.kt
@@ -1,4 +1,4 @@
-package com.github.kitakkun.foos.ui.view.component.button
+package com.github.kitakkun.foos.customview.composable.button
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.BoxScope
