+++ user/src/main/java/com/github/kitakkun/foos/user/UserList.kt
@@ -1,4 +1,4 @@
-package com.github.kitakkun.foos.ui.view.component.list
+package com.github.kitakkun.foos.user
 
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
@@ -8,7 +8,6 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
 import com.github.kitakkun.foos.common.ext.OnAppearLastItem
-import com.github.kitakkun.foos.ui.state.screen.followlist.UserItemUiState
 
 @Composable
 fun UserList(
