+++ user/src/main/java/com/github/kitakkun/foos/user/followlist/FollowListScreenUiState.kt
@@ -1,4 +1,6 @@
-package com.github.kitakkun.foos.ui.state.screen.followlist
+package com.github.kitakkun.foos.user.followlist
+
+import com.github.kitakkun.foos.user.UserItemUiState
 
 data class FollowListScreenUiState(
     val followers: List<UserItemUiState>,
