+++ user/src/main/java/com/github/kitakkun/foos/user/repository/FollowRepositoryImpl.kt
@@ -1,7 +1,7 @@
-package com.github.kitakkun.foos.data.repository
+package com.github.kitakkun.foos.user.repository
 
-import com.github.kitakkun.foos.data.model.database.DatabaseFollow
-import com.github.kitakkun.foos.util.join
+import com.github.kitakkun.foos.user.join
+import com.github.kitakkun.foos.user.model.DatabaseFollow
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.firestore.FieldValue
 import com.google.firebase.firestore.FirebaseFirestore
